Rename coms to coms_debug
